requirements.txt
click==7.1.2
Flask==1.1.2
psycopg2-binary==2.8.5
gevent==20.5.0
greenlet==0.4.15
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
flask-restx==0.2.0
Flask-SQLAlchemy==2.4.1
Flask-Migrate==2.5.3
Flask-Script==2.0.6
cloudpickle==1.4.1
numpy==1.18.4
pandas==0.24.1
scikit-learn==0.20.2
sklearn==0.0
unicode==2.7
Unidecode==1.1.1
xgboost==0.90
